Scalable version of a autoencoder (conv --> deconv, all valid).
Network is goverend by parameters
filters: number of output filters at each rep operation in downward path (and in reverse for upward path)

rep: the number of repeated operations in (with the current filter size). The output of the last conv in rep is skipped through the network to its respective mirrored deconv.

kernelsize: size of the filter applied in every operation!